feat(admin): add exhibitions workspace
