+++ apps/web/app/admin/ClientPage.tsx
@@ -283,6 +283,7 @@ export default function AdminPage() {
             {
               title: 'Marketing',
               links: [
+                { href: '/admin/exhibitions', label: 'Exhibitions' },
                 { href: '/admin/voucher-codes', label: 'Voucher Management' },
                 { href: '/admin/email-schedules', label: 'Email Schedules' },
               ],
