+++ apps/web/app/admin/exhibitions/page.tsx
@@ -1,0 +1,7 @@
+import ClientPage from "./ClientPage";
+
+export const metadata = { title: "Exhibitions | Pineapple Tapped" };
+
+export default function Page() {
+  return <ClientPage />;
+}
